pilih admin

+++ order_confirm.php
@@ -6,38 +6,46 @@ if ( !($_SESSION['userid'] == 1 || $_SESSION['userid'] == 4) ) {
     die("ACCESS DENIED");
 }
 
-
-
+if (!isset($_SESSION['admin'])){
+    $_SESSION['admin'] = "";
+}
 
 $sql = 
-" SELECT 
-orders.orderid as orderid,
-orders.orderdate,
-orders.userid,
-orders.status,
-users.username,
-users.phone,
-users.address,
-item_attributes.attributeid as attributeid,
-item_attributes.size as size,
-item_attributes.weight as weight,
-item_attributes.price as price,
-items.name,
-items.image
-FROM orders
-LEFT JOIN users
-ON orders.userid = users.userid 
-LEFT JOIN item_attributes
-ON orders.attributeid = item_attributes.attributeid
-LEFT JOIN items
-ON item_attributes.itemid = items.itemid
-WHERE orders.status ='pending'
-ORDER BY orderid DESC
-";
-$stmt = $pdo->query($sql);
-$orders = $stmt->fetchALL(PDO::FETCH_ASSOC);
+        " SELECT 
+        orders.orderid as orderid,
+        orders.orderdate,
+        orders.userid,
+        orders.admin,
+        users.username,
+        users.phone,
+        users.address,
+        item_attributes.attributeid as attributeid,
+        item_attributes.size as size,
+        item_attributes.weight as weight,
+        item_attributes.price as price,
+        items.name,
+        items.image
+        FROM orders
+        LEFT JOIN users
+        ON orders.userid = users.userid 
+        LEFT JOIN item_attributes
+        ON orders.attributeid = item_attributes.attributeid
+        LEFT JOIN items
+        ON item_attributes.itemid = items.itemid
+        WHERE orders.status ='pending' AND
+        orders.admin LIKE :admin 
+        ORDER BY orderid DESC
+        ";
+        $stmt = $pdo->prepare($sql);
+        $stmt->execute(array(
+            ":admin" =>  "%".$_SESSION['admin']."%"));
+        $orders = $stmt->fetchALL(PDO::FETCH_ASSOC);
 
-
+    if(isset($_POST['cari'])){       
+        $_SESSION['admin'] = $_POST['admin'];
+        header("Location: order_confirm.php");
+        return;
+    }
 
     if(isset($_POST['action'])){
         if($_POST['action'] == 'Approve'){
@@ -160,7 +168,7 @@ $orders = $stmt->fetchALL(PDO::FETCH_ASSOC);
             }
         td:nth-of-type(1):before { content: "Orderid"; }
         td:nth-of-type(2):before { content: "Tanggal"; }
-        td:nth-of-type(3):before { content: "Status"; }
+        td:nth-of-type(3):before { content: "Admin"; }
         td:nth-of-type(4):before { content: "Userid"; }
         td:nth-of-type(5):before { content: "Nama"; }
         td:nth-of-type(6):before { content: "No WA"; }
@@ -236,12 +244,39 @@ $orders = $stmt->fetchALL(PDO::FETCH_ASSOC);
 
 <section class="container">
 
+
     <div class="box-table">
+    <div class="box">
+    <form method="post"  id="order-confirm">
+       <h1>Konfirmasi Order Web</h1>
+            <div class="form-field">
+                <label for="admin" style="font-size: 20px;">Pilih Admin :</label>
+                <select name="admin" >
+                    <option <?php if($_SESSION['admin'] === ''){echo("selected = \"selected\"");}?>value="">Semua</option>
+                    <option <?php if($_SESSION['admin'] === 'shopee'){echo("selected = \"selected\"");}?>value="shopee">Shopee</option>
+                    <option <?php if($_SESSION['admin'] === 'tokped'){echo("selected = \"selected\"");}?>value="tokped">Tokped</option>
+                    <option <?php if($_SESSION['admin'] === 'IG'){echo("selected = \"selected\"");}?>value="IG">IG</option>
+                    <option <?php if($_SESSION['admin'] === '1'){echo("selected = \"selected\"");}?>value="1">Admin 1</option>
+                    <option <?php if($_SESSION['admin'] === '2'){echo("selected = \"selected\"");}?>value="2">Admin 2</option>
+                    <option <?php if($_SESSION['admin'] === '3'){echo("selected = \"selected\"");}?>value="3">Admin 3</option>
+                    <option <?php if($_SESSION['admin'] === '4'){echo("selected = \"selected\"");}?>value="4">Admin 4</option>
+                    <option <?php if($_SESSION['admin'] === '5'){echo("selected = \"selected\"");}?>value="5">Admin 5</option>
+                <select>		
+            </div>
+            <br>
+            <div class="form-field">
+  				<input id="Submit" type="submit" name="cari" value="Go" class="button">
+  			</div>
+            <br>
+            
+       </form>
+    </div>
+
     <table>
             <tr>
               <th>OrderId</th>
               <th>OrderDate</th>
-              <th>Status</th>
+              <th>Admin</th>
               <th>Userid</th>
               <th>Username</th>
               <th>No WA</th>
@@ -257,27 +292,30 @@ $orders = $stmt->fetchALL(PDO::FETCH_ASSOC);
             </tr>
             
             <?php
-            foreach ($orders as $row) {
-                echo ("<form method=\"post\">");
-                echo ("<input type=\"hidden\" name=\"att_id\" value=\"".$row['attributeid']."\">");
-                echo ("<input type=\"hidden\" name=\"orderid\" value=\"".$row['orderid']."\">");
-                echo ("<tr>");
-                echo ("<td>".$row['orderid']."</td>");
-                echo ("<td>".$row['orderdate']."</td>");
-                echo ("<td>".$row['status']."</td>");
-                echo ("<td>".$row['userid']."</td>");
-                echo ("<td>".$row['username']."</td>");
-                echo ("<td>".$row['phone']."</td>");
-                echo ("<td>".$row['address']."</td>");
-                echo ("<td>".$row['name']."</td>");
-                echo ("<td><img class=\"logo\" src=\"item-image/".$row['image']."\"</td>");
-                echo ("<td>".$row['size']."</td>");
-                echo ("<td>".$row['weight']."</td>");
-                echo ("<td>".$row['price']."</td>");
-                echo ("<td><input type=\"submit\" name=\"action\"class=\"approve\"value=\"Approve\"onClick=\"return confirm('Approve?') \"></td><td><input type=\"submit\"name=\"action\" class=\"reject\"value=\"Reject\"onClick=\"return confirm('Reject?') \"></td>");
-                echo ("<td><input type=\"submit\" name=\"action\"class=\"delete\"value=\"Delete\"onClick=\"return confirm('Barang sudah habis?') \"></td>");
-                echo ("</tr>");
-                echo ("</form>");
+            if(isset($orders)){
+                foreach ($orders as $row) {
+                    echo ("<form method=\"post\">");
+                    echo ("<input type=\"hidden\" name=\"att_id\" value=\"".$row['attributeid']."\">");
+                    echo ("<input type=\"hidden\" name=\"orderid\" value=\"".$row['orderid']."\">");
+                    echo ("<input type=\"hidden\" name=\"admin\" value=\"".$_SESSION['admin']."\">");
+                    echo ("<tr>");
+                    echo ("<td>".$row['orderid']."</td>");
+                    echo ("<td>".$row['orderdate']."</td>");
+                    echo ("<td>".$row['admin']."</td>");
+                    echo ("<td>".$row['userid']."</td>");
+                    echo ("<td>".$row['username']."</td>");
+                    echo ("<td>".$row['phone']."</td>");
+                    echo ("<td>".$row['address']."</td>");
+                    echo ("<td>".$row['name']."</td>");
+                    echo ("<td><img class=\"logo\" src=\"item-image/".$row['image']."\"</td>");
+                    echo ("<td>".$row['size']."</td>");
+                    echo ("<td>".$row['weight']."</td>");
+                    echo ("<td>".$row['price']."</td>");
+                    echo ("<td><input type=\"submit\" name=\"action\"class=\"approve\"value=\"Approve\"onClick=\"return confirm('Approve?') \"></td><td><input type=\"submit\"name=\"action\" class=\"reject\"value=\"Reject\"onClick=\"return confirm('Reject?') \"></td>");
+                    echo ("<td><input type=\"submit\" name=\"action\"class=\"delete\"value=\"Delete\"onClick=\"return confirm('Barang sudah habis?') \"></td>");
+                    echo ("</tr>");
+                    echo ("</form>");
+                }
             }
             ?>
     </table>
